backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.2
django-modeltranslation==0.19.14
django-unfold==0.63.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
whitenoise==6.9.0
